requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
WTForms==3.1.2
werkzeug==2.3.7
python-dotenv==1.0.1
email-validator==2.1.1
gunicorn==21.2.0